convert to es6, common js modules
